+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Added `findByPids` function to `AbstractRepository`
+* Improved `AbstractRepository`
+* Added `getSubpages` and `getSubpagesPids` function to `PageService`
+* Improved `PageService`
 * Added `AbstractTagBasedViewHelper`
 * Added `AbstractViewHelper`